+ Poshuku: Clever URL line edit text parser.
Now we threat slightly differently user's input that is parsed by QUrl
as having an empty scheme().
Now if user's input contains at least one ., think of it as an address
and automatically prepend http://. If user's input contains no dots,
think of it as about a search query and go to Google.