[Overall] Replaced КиБ -> КБ, МиБ -> МБ in Russian translations.
Ok, we have gov ruling of year 2009 N 879 stating that usual decimal
prefixes should be used with bytes meaning powers of two. So, 2^20 is
really just a "Мбайт", not "Мебибайт". It's not entirely clear whether
"Мбайт" should further be abbreviated to "МБ" or "Мб", but, well, pull
requests are welcome.

This is stupid, though.

Fixes #1807.