Revert "[Util] Initial port of Oral to boost.hana."
Turns out hana is not exactly great for the kind of work that oral does.
Namely, instantiating compile-time strings look like a lot of work for
the compiler, while Fusion just creates the respective const char* stuff
directly.

We could have had some benefit from hana's strings, but there are a few
operations missing (like dropping the last character in FieldNameMorpher),
and whether it is worth moving to compile-time remains to be seen.