[LMP] Remove audio preview feature
Sadly, the internets are not the same as in 2012, nor they will ever be
like that again. There is no more anarchy like that, and, in particular,
there is no more this easily accessible free music. Thus, there is just
nothing to preview from.

Hold onto your spotifies, deezers and what not.