[Overall] Rework how tabs interface with the Core
Previously, notifying the Core about tab creation/changes was the
duty of the corresponding plugin's object via the IHaveTabs signals.
This implies all the signals from the tabs (and from the contexts where
the tabs can be created) should've been routed to the plugin object,
leading to quite centralized designs and some code duplication.

This change makes the tabs themselves notify about the changes (like tab
title or icon change), and replaces the addNewTab() signal with an
IRootWindowsManager method.

All in all, a huge refactoring that's long overdue.