[LMP] Cache the play stats instead of formatted tooltips
This simplifies updating the tooltips (indeed, updates happen naturally
now) with the passage of time.

If the tooltip text is cached, then special care must be taken to
properly invalidate relative references like "Last playback: yesterday"
as "yesterday" becomes an obsolete reference. But, actual tooltip
formatting takes about 1% of the stats collection time, so we might as
well just cache the latter and always (re)generate the former.