[Aggregator] Avoid UB due to std::views::iota() contract being invented by morons
A proper iota(s, e) returns an empty range if `s > e`, but in the C++
la-la-land it is an undefined behavior.
And, due to ranges being very heavily templated and carrying everything
in its types, one cannot just throw up a simple wrapper like
```
auto Iota(T start, T end)
{
  return start <= end ? views::iota(start, end) : views::empty<T>;
}
```
because the types are incompatible, and there doesn't seem to be an easy
way of actually making them compatible.

So here we are, writing a bunch of ugly checks.