[Util] Abort `Left`-y `Either` via a special exception
The previous approach of  `co_await`ing on promise's `final_suspend()`
(that is, `FinalSuspender` of the `Task`) had lots of problems and
didn't really work.

Firstly, `FinalSuspender::await_suspend()` never called the calling
coroutine, so the sub-`Task` spawned by `EitherAwaiter::await_suspend()`
didn't ever got control back from the `co_await`.
And that is the correct behavior, since `FinalSuspender` is supposed to
be, well, final, and the corresponding coroutine is destroyed right
after anyway.

Sure, this could have been worked around one way or another (for
example, by checking if the caller's coroutine address is the same as
the `FinalSuspender`'s coroutine address, and resuming the caller only
if they differ). But, the next problem then is that `handle->destroy()`
in `EitherAwaiter` (which, remember, never executed before) will then
get invoked and will destroy the original coroutine. That's sort of what
we want anyway, but the problem is that, for example, the `Task`
representing the coroutine might well be alive, and there's no way of
knowing that the coroutine died (and if there was, getting the result
of the coroutine out of the dead promise would be funny anyway).

All of this is avoided if an exception is thrown inside the coroutine.
Then the normal exception handling mechanism in coroutines kicks in and
does exactly what we want. The only remaining thing is to check the
exception type in `GetTaskResult()` and ignore that particular exception
type.

So, yeah, that's sorta exceptions for control flow. But it is what it
is.