[Monocle] Use vertical scrollbar size from the pixel metric
The actual scroll bar might not be visible (and thus it might report
invalid size), so just use the `QStyle`'s pixel metric.

Sure, this means that we might be overly pessimistic and spend a few
extra pixels on the scrollbar when it's not gonna be visible, but such
situations will be extremely rare (most documents don't fit on the
screen), so it's a reasonable compromise.