[Util] Resume the coroutine immediately when its context object dies
This way, when the context object dies, we can immediately throw the
corresponding exception (when the coroutine `resume()`s and goes through
the `AwaitableWrapper::await_resume()` wrapping the current awaitable)
instead of waiting until the current awaitable (such as a timer or a
network request) resumes it instead (when the timer is up or the request
is done).

This way, Q_OBJECT's methods can really be coroutines without worrying
about lifetimes too much, merely by adding `this` as a context object.

Indeed, before this change, if `this` is destroyed before the current
awaitable (timer/request) finishes, then the coroutine wakes up only
when the awaitable finishes some time later on, with `this` being a
dangling pointer and, if I'm reading the standard correctly, merely
waking up such a coroutine is UB.

Now, `QObject::destroyed()` causes the coroutine to `resume()`
immediately while the object is still being destroyed (and `this` being
valid), if only to throw the dead object exception and finish (abort,
really) the coroutine while `this` is still good.