[BT] Split temporary and permament alert handlers
The temp handlers need to be removed, so the removal RAII class keeps a
reference to the corresponding list. The problem is this list is in a
QHash, which might move the elements on rehashing, invalidating the
reference (see ef9d1dc1bc1d9e56c26663623e27d0e3b1f86944). So, we need to
either keep everything in a stable container (like a QMap), degrading
the lookup performance for those handlers that are never deleted,
or keep a separate QMap for temporary containers.