[XSD] Fix potential miscompilation with Qt 4.9
For some reason iterator comparison in Qt 4.9 instantiates a SFINAE-based
check for the comparison operator on the list elements, which fails if the
list element's type has an operator< (or operator<=>) that instantiates an
unavailable operator< (or operator<=>) of a type it aggregates.

The simplest fix than is to just have a type that does not define any
comparison operators at all.

As a short example, see https://gcc.godbolt.org/z/b77cYe7hE or
```
#include <QPair>
#include <QPointer>
#include <variant>
#include <functional>
#include <QList>

void foo(QList<QPair<bool, std::variant<std::function<void ()>>>> subscribers)
{
	subscribers.removeIf ([] (const auto& pair) { return !pair.first; });
}
```