[Util] Don't forward case sensitivity to QSortFilterProxyModel
The base model will never do the filtering itself, so there's no point
in synchronizing this.

In an ideal world,FixedStringFilterProxyModel wouldn't even derive from
QSortFilterProxyModel, but so far this is the only way to avoid
reimplementing the non-trivial and error-prone row filtering/removal
logic (irrespective of the specific filtering predicate) ourselves.